backend-flask/requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.5
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0

# ML/AI Dependencies
tensorflow==2.15.0
librosa==0.10.1
numpy==1.24.3
scipy==1.11.4

# Additional utilities
python-multipart==0.0.6
